+++ src/components/client/UserDetails/components/DetailsPassword.tsx
@@ -3,17 +3,14 @@
 import cn from "@/utils/cn";
 import { Details, useDetailsForm } from "@/components/client/UserDetails";
 import { BaseTL } from "@/app/[locale]/(auth)/account/page";
+import { useTranslations } from "next-intl";
 
 type DetailsPasswordProps = {
   canEdit: boolean;
-  tl: Omit<BaseTL, "placeholder"> & {
-    currPas: string;
-    newPass: string;
-    repeatPass: string;
-  };
 };
 
-export function DetailsPassword({ canEdit, tl }: DetailsPasswordProps) {
+export function DetailsPassword({ canEdit }: DetailsPasswordProps) {
+  const t = useTranslations("Account");
   const { isEditing, setIsEditing, formAction, error } = useDetailsForm("");
 
   return (
@@ -21,11 +18,11 @@ export function DetailsPassword({ canEdit, tl }: DetailsPasswordProps) {
       <Details.View
         value="*****"
         canEdit={canEdit}
-        title={tl.title}
+        title={t("your_password")}
         onClick={() => setIsEditing(true)}
       >
         <Details.Overlay
-          title={tl.changeTitle}
+          title={t("change_password")}
           isEditing={isEditing}
           setIsEditing={setIsEditing}
         >
@@ -33,7 +30,7 @@ export function DetailsPassword({ canEdit, tl }: DetailsPasswordProps) {
             <Details.Input
               type="password"
               name="currentPass"
-              placeholder={tl.currPas}
+              placeholder={t("password_current")}
               className={cn(error.password?.old && "border-2 border-red-500")}
             />
             {error.password?.old && (
@@ -44,13 +41,13 @@ export function DetailsPassword({ canEdit, tl }: DetailsPasswordProps) {
             <Details.Input
               type="password"
               name="newPass"
-              placeholder={tl.newPass}
+              placeholder={t("password_new")}
               className={cn(error.password?.new && "border-2 border-red-500")}
             />
             <Details.Input
               type="password"
               name="repeatedPass"
-              placeholder={tl.repeatPass}
+              placeholder={t("password_repeat")}
               className={cn(error.password?.new && "border-2 border-red-500")}
             />
             {error.password?.new && (
@@ -58,7 +55,7 @@ export function DetailsPassword({ canEdit, tl }: DetailsPasswordProps) {
                 {error.password?.new}
               </span>
             )}
-            <Details.Submit tl={tl.submitTL} isEditable />
+            <Details.Submit isEditable />
           </form>
         </Details.Overlay>
       </Details.View>
