+++ src/components/client/UserDetails/components/common/DetailsSubmit.tsx
@@ -3,24 +3,20 @@
 import { useFormStatus } from "react-dom";
 
 import Button from "@/components/common/Button";
-
-export type DetailsSubmitTL = {
-  editable: string;
-  notEditable: string;
-};
+import { useTranslations } from "next-intl";
 
 type DetailsSubmitProps = {
   isEditable: boolean;
-  tl: DetailsSubmitTL;
 };
 
-export function DetailsSubmit({ isEditable, tl }: DetailsSubmitProps) {
+export function DetailsSubmit({ isEditable }: DetailsSubmitProps) {
   const { pending } = useFormStatus();
+  const t = useTranslations("Account");
 
   return (
     <div className="grid grid-cols-1 gap-2 font-semibold disabled:[&_button]:cursor-no-drop">
       <Button isSubmit disabled={!isEditable || pending}>
-        {isEditable ? tl.editable : tl.notEditable}
+        {isEditable ? t("submit_allowed") : t("submit_disallowed")}
       </Button>
     </div>
   );
